+++ app/assets/urls/riseup.url
@@ -1,4 +1,4 @@
 {
 	"main_url" : "https://riseup.net/",
-	"ca_cert_fingerprint" : "c429750763322fcc20b6402d7b508ba9e754b176"
+	"ca_cert_fingerprint" : "aef7a642d7f8e046770521b354961a95cd4a76a8"
 }
