+++ app/build.gradle
@@ -1,5 +1,4 @@
-apply plugin: 'android'
-import org.ajoberstar.grgit.Grgit
+apply plugin: 'com.android.application'
 
 android {
   compileSdkVersion 21
@@ -14,6 +13,15 @@ android {
     }
   }
 
+    productFlavors {
+        production {
+
+        }
+        insecure {
+
+        }
+    }
+
   buildTypes {
     release {
       //runProguard true
@@ -22,9 +30,9 @@ android {
     }
   }
 
-  lintOptions {
-    abortOnError false
-  }
+    lintOptions {
+        abortOnError false
+    }
 
   sourceSets {
     main {
