--- app/src/debug/assets/urls/dev.bitmask.net.url
@@ -1,3 +1,0 @@
-{
-	"main_url" : "https://dev.bitmask.net/"
-}
