+++ app/src/insecure/AndroidManifest.xml
