+++ app/src/insecure/assets/urls/dev.bitmask.net.url
@@ -1,0 +1,3 @@
+{
+	"main_url" : "https://dev.bitmask.net/"
+}
