+++ app/src/insecure/java/se/leap/bitmaskclient/NewProviderDialog.java
