+++ app/src/insecure/res/layout-xlarge/new_provider_dialog.xml
