+++ app/src/insecure/res/layout/new_provider_dialog.xml
