+++ app/src/production/java/se/leap/bitmaskclient/NewProviderDialog.java
