+++ app/src/production/res/layout-xlarge/new_provider_dialog.xml
