+++ app/src/production/res/layout/new_provider_dialog.xml
